Pipelines for all needed types of LOC content??
Collections, items, API queries, and newspapers benefit from slightly
different handling. I've created that for each of them, as well as
pulled together lists of potentially useful items, queries, etc. from
various sources (Eric Foner's bibliography, speaking to LOC experts).

The expectation is that the content returned here will be overbroad, but
it will be easier to figure out *how* it's overbroad (hence how to
narrow it) once I'm able to visualize what's here.

Git committing this to make it easier to deploy off of localhost, so
that I can have lots of space to pull down content (the ChronAm in
particular is going to take up a few hundred GB). I'll do a run there
and look for:
- are there other exceptions I need to handle
- are there additional subject headers suggested by what I encounter
  commonly here (initial investigation suggests no -- overwhelmingly I
  get geographic terms like "united states")

I don't have any way of checkpointing this right now, which is a bummer,
as it takes a very long time to run to completion. Fingers crossed!