Fix column header
Now I can visualize this with quadfeather/deepscatter! The default
params are all wrong, and without hovercards I can't tell what I'm
looking at, but I can SEE it and that is so cool.

The trick was making sure that the column headers referenced in the
"encoding" prefs matched the columns that actually exist in the csv.