Update metadata handling pipeline
If we load all the metadata into memory at once it's too big, lol.
Write it one item at a time and handle it one item at a time. We're
using the filesystem as a discount database here -- file handles (which
are LOC identifiers) as primary keys, because it's easier than standing
up infrastructure.