Bugfix place name filtration
We lowercased the words to tokenize them for training, but we never
actually changed the case in the files. Therefore this filter was not
in general matching things, because place names are capitalized!

Instead, let's make sure to normalize the text in the same way that we
normalize the stopwords.