Don't pipenv install numpy/scipy on M1
I'd actually rather a "!= darwin" to "==linux", but I don't think that's
a thing, and I'm only using M1 and linux.

All of these deps are still needed, but must be manually installed.

I hear poetry deals with this seamlessly (because it can speak conda,
and all these install paths work on conda) but switching hasn't been a
high enough priority.