Add cache to nonword filter
It's about 3% faster, on a sample of ~3500 documents, to have the cache.
It's slightly slower than the non-cached version on the first few
hundred documents, presumably due to overhead in setting up and checking
the cache. However, the relative performance improves over time as the
cache is built, so it should be a noticeable advantage on a large
corpus.