Parallelize nonword filter
The filter to replace nonwords is, necessarily, extremely slow, since
it must look at every word in every document. However, we can
parallelize it! Yay!

Parallelizing is slower for very small corpora (there's some overhead
in, e.g., needing to load the model for every thread, and frequent
db commits), but it's twice as fast on a sample corpus of ~3600
newspapers, using 4 processes, and I suspect the gap grows with corpus
size.