Test with lots of filters
This test looks to...actually be putting things together more
conceptually?? Of course there are still visible collection boundaries
-- e.g. Elizabeth Cady Stanton's papers are generally together, being
generally about women's rights -- but they are near Susan B. Anthony's
papers, AND near https://www.loc.gov/item/rbpe.16000300/, a declaration
on women's rights. A Walt Whitman poem is near a newspaper that was
running a poem.

I'm writing metadata wrong because I'm not fully processing newspaper
directories (the code needs to be more agnostic about the containing
directories for newspaper files) but this is starting to look meaningful
in the test visualization!