Add tests of filters
I might want to rewrite the pipeline in future so that filters read from
one location and write to a second one -- that would make it easier to
keep track of stages in a pipeline. It'll be straightforward to add that
to these tests, though, and I'm happier having the tests to start with.